+ Move static exception defaults to ExceptionHandler + Move `stringifyErrorSafely()` to `ExceptionHandler.stringifyErrorSafely()`
