+++ exception/index.js
@@ -5,6 +5,5 @@
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 export * from './Exception';
-export * from './Stringify';
